warnings from WDK 6000
fixed a few cases of warning 4242: possible loss of data

wincrypt.h appears to come with windows.h in later versions of the ddk, but nfs41_client.c fails to compile in WDK 6001 without #include <wincrypt.h>

+++ daemon/pnfs.h
@@ -316,7 +316,7 @@ __inline uint32_t stripe_index(
     IN uint64_t sui,
     IN uint32_t stripe_count)
 {
-    return (sui + layout->first_index) % stripe_count;
+    return (uint32_t)((sui + layout->first_index) % stripe_count);
 }
 __inline uint32_t data_server_index(
     IN pnfs_file_device *device,
